+++ src/components/Chat/ChattingHandler.tsx
@@ -8,7 +8,7 @@ import signGame from "@/components/api/SignCreate";
 import gameAPI from "@/components/api/Game";
 import ChatInput from "./ChatInput";
 import ChatMessage from "./ChatMessage";
-
+import { useNavigate } from "react-router-dom";
 
 interface Chatting {
   externalId?: string | null;
@@ -20,11 +20,6 @@ interface Chatting {
   data?: any | null;
 }
 
-interface ChattingComponentProps {
-  homeInputText: string;
-  resetInput: () => void;
-  changeParentsFunction: () => void;
-}
 
 interface GameRelation {
   key: string;
@@ -51,11 +46,7 @@ interface ChatMessage {
   data: GameData;
 }
 
-function ChattingComponent({
-  homeInputText,
-  resetInput,
-  changeParentsFunction,
-}: ChattingComponentProps) {
+function ChattingComponent() {
   const { user } = usePrivy();
   const twitterAccount = user?.linkedAccounts[0] as
     | { username: string }
@@ -114,9 +105,9 @@ There are three options you can choose from:
   const marketOptions = useMemo(() => {
     return messages.filter((msg) => msg.messageType === "MARKET_OPTIONS");
   }, [messages]);
-  console.log("messages",messages)
-  console.log("marketOptions",marketOptions)
-
+  console.log("messages", messages);
+  console.log("marketOptions", marketOptions);
+  const navigate = useNavigate();
   const { wallets } = useSolanaWallets();
   const wallet = wallets.find((w) => w.walletClientType === "privy");
   const [selectedChoice, setSelectedChoice] = useState<string | null>(null);
@@ -127,21 +118,6 @@ There are three options you can choose from:
   const [loading, setLoading] = useState<boolean>(false);
   const [externalId, setExternalId] = useState<string | null>(null);
 
-  useEffect(() => {
-    if (homeInputText.trim()) {
-      // const newMessage: Chatting = {
-      //   externalId: null,
-      //   content: homeInputText,
-      //   messageType: "TEXT",
-      //   sender: null,
-      // };
-      // setMessages((prevMessages) => [...prevMessages, newMessage]);
-      // setPrevHomeInputText(homeInputText);
-      // homeSendMessage(newMessage);
-
-      resetInput();
-    }
-  }, [homeInputText, resetInput]);
 
   // Scroll when the message list is updated
   useEffect(() => {
@@ -263,8 +239,6 @@ There are three options you can choose from:
     } catch (error) {
       console.error("Error sending message:", error);
     } finally {
-      changeParentsFunction();
-
       setTimeout(async () => {
         try {
           await gameAPI.fetchGameHistory({
@@ -278,6 +252,7 @@ There are three options you can choose from:
       }, 1000);
       setLoading(false);
     }
+    navigate("/");
   };
 
   return (
@@ -302,22 +277,6 @@ There are three options you can choose from:
           setInputText={setInputText}
           loading={loading}
         />
-      </div>
-
-      <div className="z-100">
-        <div
-          className="peer gap-2 p-3 opacity-30 hover:opacity-100 transition-all duration-300 text-[#B3B3B3] hover:text-white flex items-center font-family font-semibold left-0 top-0 absolute cursor-pointer"
-          onClick={changeParentsFunction}
-        >
-          <img src="PrediX-logo.webp" alt="logo" className="size-8 " />
-          <p>PrediX</p>
-        </div>
-
-        <div className="absolute left-60 top-2 hidden peer-hover:block p-2 bg-[#1E1E1E] text-white rounded-md font-bold shadow-[0px_0px_30px_rgba(255,255,255,0.4)]">
-          <div className="absolute left-[-10px] top-1/2 transform -translate-y-1/2 w-0 h-0 border-t-[10px] border-b-[10px] border-r-[10px] border-transparent border-r-[#1E1E1E]"></div>
-          Back home
-        </div>
-
       </div>
     </div>
   );
