+++ src/components/wallet/CopyQRClipboard.tsx
