+++ src/page/chat.tsx
@@ -1,0 +1,11 @@
+import ChattingComponent from "@/components/chat/ChattingHandler";
+
+function chat() {
+  return (
+    <div>
+      <ChattingComponent/>
+    </div>
+  );
+}
+
+export default chat;
